+++ NEWS.md
@@ -1,3 +1,14 @@
+# jsTreeR 2.0.0 (2022-06-08)
+
+- The package now provides the 'tree navigator' Shiny module, which allows to 
+render a files and folders navigator in the server side file system.
+
+- New Shiny input value accessible in `input$ID_selected_tree`. This is like 
+`input$ID_selected` but it also provides the ascendants of the selected nodes. 
+
+- Upgraded 'jsTree' library to the development version.
+
+
 # jsTreeR 1.6.0 (2022-02-28)
 
 New Shiny input value accessible in `input$ID_selected_paths`. This is like